Move HttpMetrics into a shared lib
